initial commit for making filter, sort, and limit functions in cache/sample general
